Improve tm-edit-tiddler, add tm-navigate-{prev,next}
In order to be usable as a keyboard shortcut, tm-edit-tiddler must work
without event parameters. In this case, default to editing the history
list's current-tiddler (not to be confused with the <<currentTiddler>>
used for rendering).

Add messages tm-navigate-prev and tm-navigate-next which navigate to the
previous/next element in the story list (with respect to the history
list's current-tiddler).